pyproject.toml
[project]
name = "thumbforge"
version = "0.1.0"
description = "Batch thumbnail generator with Krita template support and AI background generation"
requires-python = ">=3.10"
dependencies = [
    "PySide6",
    "Pillow",
    "lxml",
]

[project.optional-dependencies]
ai = [
    "requests",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
commit_parser = "conventional"
tag_format = "v{version}"
allow_zero_version = true
major_on_zero = true

[tool.semantic_release.changelog]
template_dir = "templates"
mode = "init"
exclude_commit_patterns = [
  "^chore\\(release\\):",
  "^Merge ",
]

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "refactor", "ci", "chore"]
parse_squash_commits = true
ignore_merge_commits = true
